requirements.txt
numpy==1.26.4
matplotlib==3.9.0
torch==2.3.0
torchvision==0.18.0
opencv-python==4.9.0.80
pillow=10.3.0
pytest=8.2.1
